refactor: Compilation to stand alone class
